move Button to common folder
