--- src/components/Button/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Button';
-export { default as NavButton } from './NavButton';
